Add files via upload
Added all the files currently created for the checkers game. These files simulate the game of checkers by the player for the coordinates of the piece they would like to move, followed by the coordinates of where they want to move that piece. It does this for each player until a win condition has been met.